+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/ordemservico/dashboard/DashboardChecklistOsResource.java
@@ -1,5 +1,6 @@
 package br.com.zalf.prolog.webservice.frota.checklist.ordemservico.dashboard;
 
+import br.com.zalf.prolog.webservice.commons.util.Required;
 import br.com.zalf.prolog.webservice.dashboard.components.charts.pie.PieChartComponent;
 import br.com.zalf.prolog.webservice.dashboard.components.table.TableComponent;
 import br.com.zalf.prolog.webservice.errorhandling.exception.ProLogException;
@@ -38,5 +39,12 @@ public final class DashboardChecklistOsResource {
             @PathParam("codComponente") final Integer codComponente,
             @QueryParam("codUnidades") final List<Long> codUnidades) throws ProLogException {
         return service.getPlacasMaiorQtdItensOsAbertos(codComponente, codUnidades);
+    }
+    @GET
+    @Path("/placas-bloqueadas/{codComponente}")
+    public TableComponent getPlacasBloqueadas(
+            @PathParam("codComponente") @Required final Integer codComponente,
+            @QueryParam("codUnidades") @Required final List<Long> codUnidades) throws ProLogException {
+        return service.getPlacasBloqueadas(codComponente, codUnidades);
     }
 }